+++ src/main/resources/db/migration/V1.0__createDB.sql
@@ -11,7 +11,7 @@ create table user_data (
     user_id varchar(255) not null,
     role varchar (50) not null,
     card_packs integer not null check (card_packs>=0),
-    coins integer not null check (coins>=0),
+    coins DOUBLE not null check (coins>=0),
     primary key (user_id));
 
 alter table card_copy add constraint FKtco9dei78cocpwi1sxye9mw3b foreign key (user_user_id) references user_data;
